+++ src/shared/preferences/index.ts
@@ -1,4 +1,5 @@
 export * from "./appearance";
 export * from "./accentColor";
+export * from "./backupStatus";
 export * from "./profile";
 export * from "./routineNudges";